Fix the sign of the constant part of a gep offset

+++ llvm/include/llvm/Cheerp/Writer.h
@@ -472,7 +472,7 @@ private:
 		{
 		}
 		void addValue(const llvm::Value* v, uint32_t size) override;
-		void addConst(uint32_t v) override;
+		void addConst(int64_t v) override;
 	};
 public:
 	ostream_proxy stream;
